filter: refactor usage
