+++ lib/indexer/filterindexer.js
@@ -10,8 +10,7 @@ const bdb = require('bdb');
 const assert = require('bsert');
 const Indexer = require('./indexer');
 const consensus = require('../protocol/consensus');
-const Filter = require('../primitives/filters');
-const GCSFilter = require('golomb');
+const Filter = require('../primitives/filter');
 
 /**
  * FilterIndexer
@@ -42,7 +41,9 @@ class FilterIndexer extends Indexer {
     const prevHash = this.network.genesis.prevBlock;
 
     // Genesis prev filter headers are defined to be zero hashes
-    await this.blocks.writeFilter(prevHash, consensus.ZERO_HASH);
+    const filter = new Filter();
+    filter.header = consensus.ZERO_HASH;
+    await this.blocks.writeFilter(prevHash, filter.toRaw());
 
     await super.saveGenesis();
   }
@@ -58,12 +59,11 @@ class FilterIndexer extends Indexer {
   async indexBlock(meta, block, view) {
     const hash = block.hash();
     const prev = await this.getFilterHeader(block.prevBlock);
-    const basic = GCSFilter.fromBlock(block, view);
+    const basic = block.toFilter(view);
 
-    const filter = Filter.fromOptions({
-      'header': basic.header(prev),
-      'filter': basic.toRaw()
-    });
+    const filter = new Filter();
+    filter.header = basic.header(prev);
+    filter.filter = basic.toRaw();
     await this.blocks.writeFilter(hash, filter.toRaw());
   }
