+++ lib/primitives/block.js
@@ -841,6 +841,11 @@ class Block extends AbstractBlock {
     return obj instanceof Block;
   }
 
+  /*
+   * Get compact block filter (BIP158)
+   * @param {CoinView} view
+   * @returns {Object} See {@link Golomb}
+   */
   toFilter(view) {
     const hash = this.hash();
     const key = hash.slice(0, 16);
