+++ lib/primitives/filter.js
